UGC and George feedbacks were modified and added files.

+++ promotionservice/src/main/java/com/kelloggs/promotions/promotionservice/service/impl/PromotionClusterServiceImpl.java
@@ -1,6 +1,8 @@
 package com.kelloggs.promotions.promotionservice.service.impl;
 
 
+import java.time.LocalDateTime;
+import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
@@ -37,26 +39,37 @@ public class PromotionClusterServiceImpl implements PromotionClusterService {
         this.promotionRepo = promotionRepo;
         this.winnerConfigRepo = winnerConfigRepo;
     }
-
+    
+    /**
+     * Add new promotion cluster
+     * 
+     * @param promotionCluster	includes new cluster name to be created
+     * @return	Return cluster entry details , otherwise throw exception if cluster already exists
+     * 
+     * @author Pranit (10715288)
+     * 
+     */
 	@Override
 	public PromotionClusterDTO addPromotionCluster(PromotionClusterDTO promotionCluster) {
-		PromotionCluster pc=null;
-		PromotionClusterDTO finalPcDTO=null;
-		if (!promotionCluster.getClusterName().isEmpty()) { // It will throw error if cluster name is empty
-			Optional<PromotionCluster> savedPromCluster=promotionClusterRepo.findByClusterName(promotionCluster.getClusterName());
-			if(savedPromCluster.isPresent()) { // It will throw the error if cluster name is already present
-				throw new ApiException(HttpStatus.BAD_REQUEST,
-						400, String.format("Cluster name is already exists in database "+ promotionCluster.getClusterName()));
+		PromotionCluster pc = null;
+		PromotionClusterDTO finalPcDTO = null;
+		if (promotionCluster.getClusterName() == null || promotionCluster.getClusterName().equals(null)
+				|| promotionCluster.getClusterName().isEmpty()) { // It will throw error if cluster name is empty
+			throw new ApiException(HttpStatus.BAD_REQUEST, 400, String.format("Cluster Name Not Found"));
+
+		} else {
+
+			Optional<PromotionCluster> savedPromCluster = promotionClusterRepo
+					.findByClusterName(promotionCluster.getClusterName());
+			if (savedPromCluster.isPresent()) { // It will throw the error if cluster name is already present
+				throw new ApiException(HttpStatus.BAD_REQUEST, 400, String
+						.format("Cluster name is already exists in database " + promotionCluster.getClusterName()));
 			}
-			pc= promotionClusterRepo.save(getPromotioCluster(promotionCluster));
-			finalPcDTO=getPromotioClusterDTO(pc,null);
-			
-		}else {
-			 throw new ApiException(HttpStatus.BAD_REQUEST,
-						400, String.format("Cluster Name Not Found"));
+			pc = promotionClusterRepo.save(getPromotioCluster(promotionCluster));
+			finalPcDTO = getPromotioClusterDTO(pc, null);
 		}
 		return finalPcDTO;
-		}
+	}
 	
 	PromotionCluster getPromotioCluster(PromotionClusterDTO promotionCluster) {
 		PromotionCluster pc=new PromotionCluster();
@@ -72,7 +85,15 @@ public class PromotionClusterServiceImpl implements PromotionClusterService {
 		return pc;
 	}
 	
-
+	/**
+     * Delete promotion cluster
+     * 
+     * @param clusterId	includes clusterId to be deleted
+     * @return	Return deleted cluster details , otherwise throw exception if cluster not present
+     * 
+     * @author Pranit (10715288)
+     * 
+     */
 	@Override
 	public PromotionClusterDTO deletePromotionCluster(Integer clusterId) {
 		Optional<PromotionCluster> pc=promotionClusterRepo.findById(clusterId);
@@ -80,16 +101,47 @@ public class PromotionClusterServiceImpl implements PromotionClusterService {
 			Optional<List<Promotion>> optionalPromotionList=promotionRepo.findByPromotionClusterId(clusterId);
 			if(optionalPromotionList.isPresent()) { // It is used to check promotions are linked to cluster id
 				for(Promotion p:optionalPromotionList.get()) {
-					Query query = entityManager.createNativeQuery(
-							"select config_id from winner_selection_config_reference where promotion_id="
-									+ p.getId());
-					List<Integer> resultList = (List<Integer>) query.getResultList();
-					if (!resultList.isEmpty()) {
-						for (Integer configId : resultList) {
-							winnerConfigRepo.deleteById(configId);
+					
+					Date startDate = p.getStartDate();
+					Date endDate = p.getEndDate();
+					String startDateInstring = startDate.toString();
+					String endDateInString = endDate.toString();
+					String currentDateInString = LocalDateTime.now().toLocalDate().toString();
+					if (startDate.toString().equals(currentDateInString) // start date is equals today
+							|| (!PromotionServiceImpl.isDatePastTodayFuture(startDateInstring + "T00:00:00.000Z")
+									&& endDateInString.equals(currentDateInString))// startdate yesterday and end date today
+							|| (!PromotionServiceImpl.isDatePastTodayFuture(startDateInstring + "T00:00:00.000Z")
+									&& (endDate.getTime() > new Date().getTime())) // start date is past date and end date today or 
+																					// is future date
+							|| (startDateInstring.equals(currentDateInString)
+									&& endDateInString.equals(currentDateInString))// start date and end date is today
+
+							|| (!PromotionServiceImpl.isDatePastTodayFuture(startDateInstring + "T00:00:00.000Z")
+									&& !PromotionServiceImpl.isDatePastTodayFuture(endDateInString + "T00:00:00.000Z"))
+							) {
+						throw new ApiException(HttpStatus.BAD_REQUEST, 400, String.format(
+								"Cluster with PromotionIds are already live now. please provide another cluster id."
+								));
+					}else {
+					
+						try {
+							Query query = entityManager.createNativeQuery(
+									"select config_id from winner_selection_config_reference where promotion_id="
+											+ p.getId());
+							List<Integer> resultList = (List<Integer>) query.getResultList();
+							if (!resultList.isEmpty()) {
+								for (Integer configId : resultList) {
+									winnerConfigRepo.deleteById(configId);
+								}
+							}
+							promotionRepo.deleteById(p.getId());
+							
+						} catch (Exception e) {
+							e.printStackTrace();
+							throw new ApiException(HttpStatus.BAD_REQUEST, 400,
+									String.format("Database error occured while deleting promotion id :"+p.getId()));
 						}
-					}
-					promotionRepo.deleteById(p.getId());
+				 }
 				}
 			}
 			promotionClusterRepo.deleteById(clusterId);
